main file resolution as tools.js method
- closes #789
