+++ lib/cli/cliDebug.js
@@ -2,8 +2,6 @@
 const CLI = require('./messages.js');
 const CLICommand = require('./cliCommand.js');
 const tools = require('../tools.js');
-const fs = require('fs');
-const path = require('path');
 const child_process = require('child_process');
 
 /** Command ioBroker debug ... */
@@ -17,7 +15,7 @@ module.exports = class CLICompact extends CLICommand {
      * Executes a command
      * @param {any[]} args
      */
-    execute(args) {
+    async execute(args) {
         const { callback, ...params } = this.options;
         const adapter = args[0];
         if (!adapter) {
@@ -34,69 +32,30 @@ module.exports = class CLICompact extends CLICommand {
             return void callback(34);
         }
 
-        // Compile a list of all possible main files
-        const possibleMainFiles = [
-            'main.js',
-            `${adapter}.js`
+        let mainFile;
+        try {
+            mainFile = await tools.getAdapterMainFile(adapter);
+        } catch (e) {
+            CLI.error.mainFileNotFound(adapter);
+            return callback(1);
+        }
+
+        // Start the adapter with force and console logs
+        const adapterArgs = ['--debug'];
+        // Tell node to attach a debugger
+        const nodeArgs = [
+            // --inspect[-brk][=[ip]:[port]]
+            `--inspect${params.wait ? '-brk' : ''}${
+                !!params.ip || !!params.port ? '=' : ''
+            }${params.ip || ''}${
+                !!params.ip && !!params.port ? ':' : ''
+            }${params.port || ''}`
         ];
-
-        // Add package.json -> main as the 2nd choice
-        try {
-            const pack = JSON.parse(
-                fs.readFileSync(
-                    path.join(adapterDir, 'package.json'),
-                    'utf8'
-                )
-            );
-            if (pack && typeof pack.main === 'string') {
-                possibleMainFiles.unshift(pack.main);
-            }
-        } catch (e) {
-            // Ignore, we have fallback solutions
-        }
-
-        // Add io-package.json -> common.main as the preferred choice
-        try {
-            const ioPack = JSON.parse(
-                fs.readFileSync(
-                    path.join(adapterDir, 'io-package.json'),
-                    'utf8'
-                )
-            );
-            if (ioPack && ioPack.common && typeof ioPack.common.main === 'string') {
-                possibleMainFiles.unshift(ioPack.common.main);
-            }
-        } catch (e) {
-            // Ignore, we have fallback solutions
-        }
-
-        // Try all possible main files
-        for (const mainFile of possibleMainFiles) {
-            const fullFileName = path.join(adapterDir, mainFile);
-            if (fs.existsSync(fullFileName)) {
-                // Start the adapter with force and console logs
-                const adapterArgs = ['--debug'];
-                // Tell node to attach a debugger
-                const nodeArgs = [
-                    // --inspect[-brk][=[ip]:[port]]
-                    `--inspect${params.wait ? '-brk' : ''}${
-                        !!params.ip || !!params.port ? '=' : ''
-                    }${params.ip || ''}${
-                        !!params.ip && !!params.port ? ':' : ''
-                    }${params.port || ''}`
-                ];
-                // And wait until the sub process has finished
-                const cp = child_process.fork(fullFileName, adapterArgs, {
-                    cwd: adapterDir,
-                    execArgv: nodeArgs
-                });
-                cp.on('exit', code => callback(code || 0));
-                // We've found our main file
-                return;
-            }
-        }
-
-        CLI.error.mainFileNotFound(adapter);
-        callback(1);
+        // And wait until the sub process has finished
+        const cp = child_process.fork(mainFile, adapterArgs, {
+            cwd: adapterDir,
+            execArgv: nodeArgs
+        });
+        cp.on('exit', code => callback(code || 0));
     }
 };
