+++ lib/tools.js
@@ -2329,6 +2329,65 @@ function pipeLinewise(input, output) {
     rl.on('line', line => output.write(line + os.EOL));
 }
 
+/**
+ * Find the adaper main file as full path
+ *
+ * @memberof tools
+ * @param {string} adapter - adapter name of the adapter, e.g. hm-rpc
+ * @returns {Promise<string>}
+ */
+async function getAdapterMainFile(adapter) {
+    const adapterDir = getAdapterDir(adapter);
+    if (!adapterDir) {
+        return Promise.reject(new Error(`Could not find adapter dir of ${adapter}`));
+    }
+
+    const possibleMainFiles = [
+        'main.js',
+        `${adapter}.js`
+    ];
+
+    // Add package.json -> main as the 2nd choice
+    try {
+        const pack = JSON.parse(
+            await fs.readFile(
+                path.join(adapterDir, 'package.json'),
+                'utf8'
+            )
+        );
+        if (pack && typeof pack.main === 'string') {
+            possibleMainFiles.unshift(pack.main);
+        }
+    } catch {
+        // Ignore, we have fallback solutions
+    }
+
+    // Add io-package.json -> common.main as the preferred choice
+    try {
+        const ioPack = JSON.parse(
+            await fs.readFile(
+                path.join(adapterDir, 'io-package.json'),
+                'utf8'
+            )
+        );
+        if (ioPack && ioPack.common && typeof ioPack.common.main === 'string') {
+            possibleMainFiles.unshift(ioPack.common.main);
+        }
+    } catch {
+        // Ignore, we have fallback solutions
+    }
+
+    // Try all possible main files
+    for (const mainFile of possibleMainFiles) {
+        const fullFileName = path.join(adapterDir, mainFile);
+        if (await fs.pathExists(fullFileName)) {
+            return Promise.resolve(fullFileName);
+        }
+    }
+
+    return Promise.reject(new Error(`Could not find main file of ${adapter}`));
+}
+
 module.exports = {
     appName: getAppName(),
     createUuid,
@@ -2338,6 +2397,7 @@ module.exports = {
     findIPs,
     generateDefaultCertificates,
     getAdapterDir,
+    getAdapterMainFile,
     getInstances,
     getAllInstances,
     getAllInstancesAsync,
